validate MarshalSortedKeys offset + add focused offset test
Address review:
- reject dataStartOffset > keys[0].ValueEnd (would serialize start > end
  and silently corrupt the first index node) with an explicit error.
- pass uint64(segmentindex.HeaderSize) at the set-compactor call site so
  the argument type is explicit.
- add TestMarshalSortedKeysDataStartOffset (first node starts at a
  non-HeaderSize offset, later nodes chain from prev ValueEnd) and a
  rejection test for an out-of-range offset.

+++ adapters/repos/db/lsmkv/compactor_set.go
@@ -264,7 +264,7 @@ func (c *compactorSet) writeIndexes(f *segmentindex.SegmentFile,
 		return fmt.Errorf("unsupported secondary indexes in compactorSet")
 	}
 
-	_, err := segmentindex.MarshalSortedKeys(f.BodyWriter(), keys, segmentindex.HeaderSize)
+	_, err := segmentindex.MarshalSortedKeys(f.BodyWriter(), keys, uint64(segmentindex.HeaderSize))
 	return err
 }
